🚧 VND44D acr: split helper modules
